+++ pkg/agent/plugin/workloadattestor/k8s/k8s_test.go
@@ -1,7 +1,3 @@
-//go:build !windows
-// +build !windows
-
-// TODO: attestor is not supported on Windows yet, skip tests until issues solved
 package k8s
 
 import (
@@ -27,7 +23,6 @@ import (
 	"github.com/hashicorp/go-hclog"
 	"github.com/sigstore/cosign/pkg/cosign/bundle"
 	"github.com/sigstore/cosign/pkg/oci"
-	"github.com/spiffe/spire/pkg/agent/common/cgroups"
 	"github.com/spiffe/spire/pkg/agent/plugin/workloadattestor"
 	"github.com/spiffe/spire/pkg/agent/plugin/workloadattestor/k8s/sigstore"
 	"github.com/spiffe/spire/pkg/common/pemutil"
@@ -40,29 +35,19 @@ import (
 	"github.com/stretchr/testify/require"
 	"google.golang.org/grpc/codes"
 	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/types"
 )
 
 const (
 	pid = 123
 
 	podListFilePath           = "testdata/pod_list.json"
-	kindPodListFilePath       = "testdata/kind_pod_list.json"
 	podListNotRunningFilePath = "testdata/pod_list_not_running.json"
-
-	cgPidInPodFilePath        = "testdata/cgroups_pid_in_pod.txt"
-	cgPidInKindPodFilePath    = "testdata/cgroups_pid_in_kind_pod.txt"
-	cgInitPidInPodFilePath    = "testdata/cgroups_init_pid_in_pod.txt"
-	cgPidNotInPodFilePath     = "testdata/cgroups_pid_not_in_pod.txt"
-	cgSystemdPidInPodFilePath = "testdata/systemd_cgroups_pid_in_pod.txt"
 
 	certPath = "cert.pem"
 	keyPath  = "key.pem"
 )
 
 var (
-	pidCgroupPath = fmt.Sprintf("/proc/%v/cgroup", pid)
-
 	clientKey, _ = pemutil.ParseECPrivateKey([]byte(`-----BEGIN PRIVATE KEY-----
 MIGHAgEAMBMGByqGSM49AgEGCCqGSM49AwEHBG0wawIBAQQgNRa/6HIy0uwQe8iG
 Kz24zEvwGiIsTDPHzrLUaml1hQ6hRANCAATz6vtJYIvPM0KOqKpdDPlsOw09hZ8P
@@ -98,47 +83,6 @@ FwOGLt+I3+9beT0vo+pn9Rq0squewFYe3aJbwpkyfP2xOovQCdm4PC8y
 		{Type: "k8s", Value: "sa:default"},
 	}
 
-	testKindPodSelectors = []*common.Selector{
-		{Type: "k8s", Value: "container-image:gcr.io/spiffe-io/spire-agent:0.8.1"},
-		{Type: "k8s", Value: "container-image:gcr.io/spiffe-io/spire-agent@sha256:1e4c481d76e9ecbd3d8684891e0e46aa021a30920ca04936e1fdcc552747d941"},
-		{Type: "k8s", Value: "container-name:workload-api-client"},
-		{Type: "k8s", Value: "node-name:kind-control-plane"},
-		{Type: "k8s", Value: "ns:default"},
-		{Type: "k8s", Value: "pod-image-count:1"},
-		{Type: "k8s", Value: "pod-image:gcr.io/spiffe-io/spire-agent:0.8.1"},
-		{Type: "k8s", Value: "pod-image:gcr.io/spiffe-io/spire-agent@sha256:1e4c481d76e9ecbd3d8684891e0e46aa021a30920ca04936e1fdcc552747d941"},
-		{Type: "k8s", Value: "pod-init-image-count:0"},
-		{Type: "k8s", Value: "pod-label:app:sample-workload"},
-		{Type: "k8s", Value: "pod-label:pod-template-hash:6658cb9566"},
-		{Type: "k8s", Value: "pod-name:sample-workload-6658cb9566-5n4b4"},
-		{Type: "k8s", Value: "pod-owner-uid:ReplicaSet:349d135e-3781-43e3-bc25-c900aedf1d0c"},
-		{Type: "k8s", Value: "pod-owner:ReplicaSet:sample-workload-6658cb9566"},
-		{Type: "k8s", Value: "pod-uid:a2830d0d-b0f0-4ff0-81b5-0ee4e299cf80"},
-		{Type: "k8s", Value: "sa:default"},
-	}
-
-	testInitPodSelectors = []*common.Selector{
-		{Type: "k8s", Value: "container-image:docker-pullable://quay.io/coreos/flannel@sha256:1b401bf0c30bada9a539389c3be652b58fe38463361edf488e6543c8761d4970"},
-		{Type: "k8s", Value: "container-image:quay.io/coreos/flannel:v0.9.0-amd64"},
-		{Type: "k8s", Value: "container-name:install-cni"},
-		{Type: "k8s", Value: "node-name:k8s-node-1"},
-		{Type: "k8s", Value: "ns:kube-system"},
-		{Type: "k8s", Value: "pod-image-count:1"},
-		{Type: "k8s", Value: "pod-image:docker-pullable://quay.io/coreos/flannel@sha256:1b401bf0c30bada9a539389c3be652b58fe38463361edf488e6543c8761d4970"},
-		{Type: "k8s", Value: "pod-image:quay.io/coreos/flannel:v0.9.0-amd64"},
-		{Type: "k8s", Value: "pod-init-image-count:1"},
-		{Type: "k8s", Value: "pod-init-image:docker-pullable://quay.io/coreos/flannel@sha256:1b401bf0c30bada9a539389c3be652b58fe38463361edf488e6543c8761d4970"},
-		{Type: "k8s", Value: "pod-init-image:quay.io/coreos/flannel:v0.9.0-amd64"},
-		{Type: "k8s", Value: "pod-label:app:flannel"},
-		{Type: "k8s", Value: "pod-label:controller-revision-hash:1846323910"},
-		{Type: "k8s", Value: "pod-label:pod-template-generation:1"},
-		{Type: "k8s", Value: "pod-label:tier:node"},
-		{Type: "k8s", Value: "pod-name:kube-flannel-ds-gp1g9"},
-		{Type: "k8s", Value: "pod-owner-uid:DaemonSet:2f0350fc-b29d-11e7-9350-020968147796"},
-		{Type: "k8s", Value: "pod-owner:DaemonSet:kube-flannel-ds"},
-		{Type: "k8s", Value: "pod-uid:d488cae9-b2a0-11e7-9350-020968147796"},
-		{Type: "k8s", Value: "sa:flannel"},
-	}
 	testSigstoreSelectors = []*common.Selector{
 		{Type: "k8s", Value: "container-image:docker-pullable://localhost/spiffe/blog@sha256:0cfdaced91cb46dd7af48309799a3c351e4ca2d5e1ee9737ca0cbd932cb79898"},
 		{Type: "k8s", Value: "container-image:localhost/spiffe/blog:latest"},
@@ -208,6 +152,7 @@ type Suite struct {
 	kubeletCert *x509.Certificate
 	clientCert  *x509.Certificate
 
+	oc                          *osConfig
 	sigstoreSelectors           []sigstore.SelectorsFromSignatures
 	sigstoreSigs                []oci.Signature
 	sigstoreSkipSigs            bool
@@ -231,6 +176,7 @@ func (s *Suite) SetupTest() {
 	s.sigstoreReturnError = nil
 	s.sigstoreSkipSigs = false
 	s.sigstoreSkippedSigSelectors = nil
+	s.oc = createOSConfig()
 }
 
 func (s *Suite) TearDownTest() {
@@ -296,27 +242,6 @@ func (s *Suite) TestAttestWithFailedSigstoreSignatures() {
 	s.Require().Contains(buf.String(), "sigstore error 123")
 }
 
-func (s *Suite) TestAttestWithPidInKindPod() {
-	s.startInsecureKubelet()
-	p := s.loadInsecurePlugin()
-
-	s.requireAttestSuccessWithKindPod(p)
-}
-
-func (s *Suite) TestAttestWithPidInPodSystemdCgroups() {
-	s.startInsecureKubelet()
-	p := s.loadInsecurePlugin()
-
-	s.requireAttestSuccessWithPodSystemdCgroups(p)
-}
-
-func (s *Suite) TestAttestWithInitPidInPod() {
-	s.startInsecureKubelet()
-	p := s.loadInsecurePlugin()
-
-	s.requireAttestSuccessWithInitPod(p)
-}
-
 func (s *Suite) TestAttestWithPidInPodAfterRetry() {
 	s.startInsecureKubelet()
 	p := s.loadInsecurePlugin()
@@ -324,7 +249,7 @@ func (s *Suite) TestAttestWithPidInPodAfterRetry() {
 	s.addPodListResponse(podListNotRunningFilePath)
 	s.addPodListResponse(podListNotRunningFilePath)
 	s.addPodListResponse(podListFilePath)
-	s.addCgroupsResponse(cgPidInPodFilePath)
+	s.addGetContainerResponsePidInPod()
 
 	resultCh := s.goAttest(p)
 
@@ -347,7 +272,7 @@ func (s *Suite) TestAttestWithPidNotInPodCancelsEarly() {
 	p := s.loadInsecurePlugin()
 
 	s.addPodListResponse(podListNotRunningFilePath)
-	s.addCgroupsResponse(cgPidInPodFilePath)
+	s.addGetContainerResponsePidInPod()
 
 	ctx, cancel := context.WithCancel(context.Background())
 	cancel()
@@ -364,7 +289,7 @@ func (s *Suite) TestAttestWithPidNotInPodAfterRetry() {
 	s.addPodListResponse(podListNotRunningFilePath)
 	s.addPodListResponse(podListNotRunningFilePath)
 	s.addPodListResponse(podListNotRunningFilePath)
-	s.addCgroupsResponse(cgPidInPodFilePath)
+	s.addGetContainerResponsePidInPod()
 
 	resultCh := s.goAttest(p)
 
@@ -386,19 +311,9 @@ func (s *Suite) TestAttestWithPidNotInPodAfterRetry() {
 	}
 }
 
-func (s *Suite) TestAttestWithPidNotInPod() {
-	s.startInsecureKubelet()
-	p := s.loadInsecurePlugin()
-	s.addCgroupsResponse(cgPidNotInPodFilePath)
-
-	selectors, err := p.Attest(context.Background(), pid)
-	s.Require().NoError(err)
-	s.Require().Empty(selectors)
-}
-
 func (s *Suite) TestAttestOverSecurePortViaTokenAuth() {
 	// start up a secure kubelet with host networking and require token auth
-	s.startSecureKubelet(true, "default-token")
+	s.startSecureKubeletWithTokenAuth(true, "default-token")
 
 	// use the service account token for auth
 	p := s.loadSecurePlugin(``)
@@ -413,7 +328,7 @@ func (s *Suite) TestAttestOverSecurePortViaTokenAuth() {
 
 func (s *Suite) TestAttestOverSecurePortViaClientAuth() {
 	// start up the secure kubelet with host networking and require client certs
-	s.startSecureKubelet(true, "")
+	s.startSecureKubeletWithClientCertAuth()
 
 	// use client certificate for auth
 	p := s.loadSecurePlugin(`
@@ -431,9 +346,19 @@ func (s *Suite) TestAttestOverSecurePortViaClientAuth() {
 	s.requireAttestFailure(p, codes.Internal, "tls: bad certificate")
 }
 
+func (s *Suite) TestAttestOverSecurePortViaAnonymousAuth() {
+	s.startSecureKubeletWithAnonymousAuth()
+
+	p := s.loadSecurePlugin(`
+		use_anonymous_authentication = true
+	`)
+
+	s.requireAttestSuccessWithPod(p)
+}
+
 func (s *Suite) TestAttestReachingKubeletViaNodeName() {
 	// start up a secure kubelet with "localhost" certificate and token auth
-	s.startSecureKubelet(false, "default-token")
+	s.startSecureKubeletWithTokenAuth(false, "default-token")
 
 	// pick up the node name from the default env value
 	s.env["MY_NODE_NAME"] = "localhost"
@@ -450,16 +375,6 @@ func (s *Suite) TestAttestReachingKubeletViaNodeName() {
 	s.requireAttestSuccessWithPod(s.loadSecurePlugin(`
 		node_name_env = "OVERRIDDEN_NODE_NAME"
 	`))
-}
-
-func (s *Suite) TestAttestAgainstNodeOverride() {
-	s.startInsecureKubelet()
-	p := s.loadInsecurePlugin()
-	s.addCgroupsResponse(cgPidNotInPodFilePath)
-
-	selectors, err := p.Attest(context.Background(), pid)
-	s.Require().NoError(err)
-	s.Require().Empty(selectors)
 }
 
 func (s *Suite) TestLogger() {
@@ -480,6 +395,9 @@ func (s *Suite) TestLogger() {
 
 func (s *Suite) TestConfigure() {
 	s.generateCerts("")
+
+	kubeletCertPool := x509.NewCertPool()
+	kubeletCertPool.AddCert(s.kubeletCert)
 
 	s.writeFile(defaultTokenPath, "default-token")
 	s.writeFile("token", "other-token")
@@ -506,8 +424,9 @@ func (s *Suite) TestConfigure() {
 		raw             string
 		hcl             string
 		config          *config
+		errCode         codes.Code
+		errMsg          string
 		sigstoreError   error
-		err             string
 		sigstoreEnabled bool
 	}{
 		{
@@ -599,9 +518,10 @@ func (s *Suite) TestConfigure() {
 		},
 
 		{
-			name: "invalid hcl",
-			hcl:  "bad",
-			err:  "unable to decode configuration",
+			name:    "invalid hcl",
+			hcl:     "bad",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to decode configuration",
 		},
 		{
 			name: "both insecure and secure ports specified",
@@ -609,21 +529,24 @@ func (s *Suite) TestConfigure() {
 				kubelet_read_only_port = 10255
 				kubelet_secure_port = 10250
 			`,
-			err: "cannot use both the read-only and secure port",
+			errCode: codes.InvalidArgument,
+			errMsg:  "cannot use both the read-only and secure port",
 		},
 		{
 			name: "non-existent kubelet ca",
 			hcl: `
 				kubelet_ca_path = "no-such-file"
 			`,
-			err: "unable to load kubelet CA",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load kubelet CA",
 		},
 		{
 			name: "bad kubelet ca",
 			hcl: `
 				kubelet_ca_path =  "bad-pem"
 			`,
-			err: "unable to parse kubelet CA",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to parse kubelet CA",
 		},
 		{
 			name: "non-existent token",
@@ -631,7 +554,8 @@ func (s *Suite) TestConfigure() {
 				skip_kubelet_verification = true
 				token_path = "no-such-file"
 			`,
-			err: "unable to load token",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load token",
 		},
 		{
 			name: "invalid poll retry interval",
@@ -639,7 +563,8 @@ func (s *Suite) TestConfigure() {
 				kubelet_read_only_port = 10255
 				poll_retry_interval = "blah"
 			`,
-			err: "unable to parse poll retry interval",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to parse poll retry interval",
 		},
 		{
 			name: "invalid reload interval",
@@ -647,7 +572,8 @@ func (s *Suite) TestConfigure() {
 				kubelet_read_only_port = 10255
 				reload_interval = "blah"
 			`,
-			err: "unable to parse reload interval",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to parse reload interval",
 		},
 		{
 			name: "cert but no key",
@@ -655,7 +581,8 @@ func (s *Suite) TestConfigure() {
 				skip_kubelet_verification = true
 				certificate_path = "cert"
 			`,
-			err: "the private key path is required with the certificate path",
+			errCode: codes.InvalidArgument,
+			errMsg:  "the private key path is required with the certificate path",
 		},
 		{
 			name: "key but no cert",
@@ -663,7 +590,8 @@ func (s *Suite) TestConfigure() {
 				skip_kubelet_verification = true
 				private_key_path = "key"
 			`,
-			err: "the certificate path is required with the private key path",
+			errCode: codes.InvalidArgument,
+			errMsg:  "the certificate path is required with the private key path",
 		},
 		{
 			name: "bad cert",
@@ -672,7 +600,8 @@ func (s *Suite) TestConfigure() {
 				certificate_path = "bad-pem"
 				private_key_path = "key.pem"
 			`,
-			err: "unable to load keypair",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load keypair",
 		},
 		{
 			name: "non-existent cert",
@@ -681,7 +610,8 @@ func (s *Suite) TestConfigure() {
 				certificate_path = "no-such-file"
 				private_key_path = "key.pem"
 			`,
-			err: "unable to load certificate",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load certificate",
 		},
 		{
 			name: "bad key",
@@ -690,7 +620,8 @@ func (s *Suite) TestConfigure() {
 				certificate_path = "cert.pem"
 				private_key_path = "bad-pem"
 			`,
-			err: "unable to load keypair",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load keypair",
 		},
 		{
 			name: "non-existent key",
@@ -699,7 +630,8 @@ func (s *Suite) TestConfigure() {
 				certificate_path = "cert.pem"
 				private_key_path = "no-such-file"
 			`,
-			err: "unable to load private key",
+			errCode: codes.InvalidArgument,
+			errMsg:  "unable to load private key",
 		},
 		{
 			name: "secure defaults with skipped images for sigstore",
@@ -777,7 +709,8 @@ func (s *Suite) TestConfigure() {
 			`,
 			sigstoreError: errors.New("rekor URL is empty"),
 			config:        nil,
-			err:           "failed to parse Rekor URL: rekor URL is empty",
+			errCode:       codes.InvalidArgument,
+			errMsg:        "failed to parse Rekor URL: rekor URL is empty",
 		},
 		{
 			name: "secure defaults for failed parsing rekor URI",
@@ -790,7 +723,8 @@ func (s *Suite) TestConfigure() {
 			`,
 			sigstoreError: errors.New("failed parsing rekor URI"),
 			config:        nil,
-			err:           "failed to parse Rekor URL: failed parsing rekor URI",
+			errCode:       codes.InvalidArgument,
+			errMsg:        "failed to parse Rekor URL: failed parsing rekor URI",
 		},
 		{
 			name: "secure defaults for invalid rekor URL Scheme",
@@ -803,7 +737,8 @@ func (s *Suite) TestConfigure() {
 			`,
 			sigstoreError: errors.New("invalid rekor URL Scheme"),
 			config:        nil,
-			err:           "failed to parse Rekor URL: invalid rekor URL Scheme",
+			errCode:       codes.InvalidArgument,
+			errMsg:        "failed to parse Rekor URL: invalid rekor URL Scheme",
 		},
 		{
 			name: "secure defaults for invalid rekor URL Host",
@@ -816,7 +751,8 @@ func (s *Suite) TestConfigure() {
 			`,
 			sigstoreError: errors.New("invalid rekor URL Host"),
 			config:        nil,
-			err:           "failed to parse Rekor URL: invalid rekor URL Host",
+			errCode:       codes.InvalidArgument,
+			errMsg:        "failed to parse Rekor URL: invalid rekor URL Host",
 		},
 	}
 
@@ -825,12 +761,14 @@ func (s *Suite) TestConfigure() {
 		s.T().Run(testCase.name, func(t *testing.T) {
 			p := s.newPlugin()
 			s.sigstoreMock.returnError = testCase.sigstoreError
+
 			var err error
 			plugintest.Load(s.T(), builtin(p), nil,
 				plugintest.Configure(testCase.hcl),
 				plugintest.CaptureConfigureError(&err))
-			if testCase.err != "" {
-				s.AssertErrorContains(err, testCase.err)
+
+			if testCase.errMsg != "" {
+				s.RequireGRPCStatusContains(err, testCase.errCode, testCase.errMsg)
 				return
 			}
 			require.NotNil(t, testCase.config, "test case missing expected config")
@@ -848,10 +786,9 @@ func (s *Suite) TestConfigure() {
 				assert.True(t, c.Client.Transport.TLSClientConfig.InsecureSkipVerify)
 				assert.Nil(t, c.Client.Transport.TLSClientConfig.VerifyPeerCertificate)
 			default:
-				t.Logf("CONFIG: %#v", c.Client.Transport.TLSClientConfig)
 				if testCase.config.HasNodeName {
 					if assert.NotNil(t, c.Client.Transport.TLSClientConfig.RootCAs) {
-						assert.Len(t, c.Client.Transport.TLSClientConfig.RootCAs.Subjects(), 1) // nolint // these pools are not system pools so the use of Subjects() is ok for now
+						assert.True(t, c.Client.Transport.TLSClientConfig.RootCAs.Equal(kubeletCertPool))
 					}
 				} else {
 					assert.True(t, c.Client.Transport.TLSClientConfig.InsecureSkipVerify)
@@ -1090,10 +1027,14 @@ func (s *Suite) kubeletPort() int {
 
 func (s *Suite) loadPlugin(configuration string) workloadattestor.WorkloadAttestor {
 	v1 := new(workloadattestor.V1)
-	plugintest.Load(s.T(), builtin(s.newPlugin()), v1,
+	p := s.newPlugin()
+	plugintest.Load(s.T(), builtin(p), v1,
 		plugintest.Configure(configuration),
 	)
 
+	if cHelper := s.oc.getContainerHelper(); cHelper != nil {
+		p.setContainerHelper(cHelper)
+	}
 	return v1
 }
 
@@ -1129,7 +1070,49 @@ func (s *Suite) generateCerts(nodeName string) {
 	s.writeCert(certPath, s.clientCert)
 }
 
-func (s *Suite) startSecureKubelet(hostNetworking bool, token string) {
+func (s *Suite) startSecureKubeletWithClientCertAuth() {
+	handler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if len(req.TLS.VerifiedChains) == 0 {
+			http.Error(w, "client auth expected but not used", http.StatusForbidden)
+			return
+		}
+		s.serveHTTP(w, req)
+	})
+
+	s.startSecureKubeletServer(false, handler)
+}
+
+func (s *Suite) startSecureKubeletWithTokenAuth(hostNetworking bool, token string) {
+	handler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if len(req.TLS.VerifiedChains) > 0 {
+			http.Error(w, "client auth not expected but used", http.StatusForbidden)
+			return
+		}
+		expectedAuth := "Bearer " + token
+		auth := req.Header.Get("Authorization")
+		if auth != expectedAuth {
+			http.Error(w, fmt.Sprintf("expected %q, got %q", expectedAuth, auth), http.StatusForbidden)
+			return
+		}
+		s.serveHTTP(w, req)
+	})
+
+	s.startSecureKubeletServer(hostNetworking, handler)
+}
+
+func (s *Suite) startSecureKubeletWithAnonymousAuth() {
+	handler := http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
+		if len(req.TLS.VerifiedChains) > 0 {
+			http.Error(w, "client auth not expected but used", http.StatusForbidden)
+			return
+		}
+		s.serveHTTP(w, req)
+	})
+
+	s.startSecureKubeletServer(false, handler)
+}
+
+func (s *Suite) startSecureKubeletServer(hostNetworking bool, handler http.Handler) {
 	// Use "localhost" in the DNS name unless we're using host networking. This
 	// allows us to use "localhost" as the host directly when configured to
 	// connect to the node name. Otherwise, we'll connect to 127.0.0.1 and
@@ -1138,32 +1121,13 @@ func (s *Suite) startSecureKubelet(hostNetworking bool, token string) {
 	if hostNetworking {
 		dnsName = "this-name-should-never-be-validated"
 	}
-	s.generateCerts(dnsName)
 
+	s.generateCerts(dnsName)
 	clientCAs := x509.NewCertPool()
 	if s.clientCert != nil {
 		clientCAs.AddCert(s.clientCert)
 	}
-	server := httptest.NewUnstartedServer(http.HandlerFunc(func(w http.ResponseWriter, req *http.Request) {
-		if token == "" {
-			if len(req.TLS.VerifiedChains) == 0 {
-				http.Error(w, "client auth expected but not used", http.StatusForbidden)
-				return
-			}
-		} else {
-			if len(req.TLS.VerifiedChains) > 0 {
-				http.Error(w, "client auth not expected but used", http.StatusForbidden)
-				return
-			}
-			expectedAuth := "Bearer " + token
-			auth := req.Header.Get("Authorization")
-			if auth != expectedAuth {
-				http.Error(w, fmt.Sprintf("expected %q, got %q", expectedAuth, auth), http.StatusForbidden)
-				return
-			}
-		}
-		s.serveHTTP(w, req)
-	}))
+	server := httptest.NewUnstartedServer(handler)
 	server.TLS = &tls.Config{
 		Certificates: []tls.Certificate{
 			{
@@ -1231,7 +1195,7 @@ func (s *Suite) writeKey(path string, key *ecdsa.PrivateKey) {
 
 func (s *Suite) requireAttestSuccessWithPod(p workloadattestor.WorkloadAttestor) {
 	s.addPodListResponse(podListFilePath)
-	s.addCgroupsResponse(cgPidInPodFilePath)
+	s.addGetContainerResponsePidInPod()
 	s.requireAttestSuccess(p, testPodSelectors)
 }
 
@@ -1245,24 +1209,6 @@ func (s *Suite) requireAttestSuccessWithPodAndSkippedImage(p workloadattestor.Wo
 	s.addPodListResponse(podListFilePath)
 	s.addCgroupsResponse(cgPidInPodFilePath)
 	s.requireAttestSuccess(p, testSigstoreSkippedSelectors)
-}
-
-func (s *Suite) requireAttestSuccessWithKindPod(p workloadattestor.WorkloadAttestor) {
-	s.addPodListResponse(kindPodListFilePath)
-	s.addCgroupsResponse(cgPidInKindPodFilePath)
-	s.requireAttestSuccess(p, testKindPodSelectors)
-}
-
-func (s *Suite) requireAttestSuccessWithPodSystemdCgroups(p workloadattestor.WorkloadAttestor) {
-	s.addPodListResponse(podListFilePath)
-	s.addCgroupsResponse(cgSystemdPidInPodFilePath)
-	s.requireAttestSuccess(p, testPodSelectors)
-}
-
-func (s *Suite) requireAttestSuccessWithInitPod(p workloadattestor.WorkloadAttestor) {
-	s.addPodListResponse(podListFilePath)
-	s.addCgroupsResponse(cgInitPidInPodFilePath)
-	s.requireAttestSuccess(p, testInitPodSelectors)
 }
 
 func (s *Suite) requireAttestSuccess(p workloadattestor.WorkloadAttestor, expectedSelectors []*common.Selector) {
@@ -1300,204 +1246,6 @@ func (s *Suite) addPodListResponse(fixturePath string) {
 	s.Require().NoError(err)
 
 	s.podList = append(s.podList, podList)
-}
-
-func (s *Suite) addCgroupsResponse(fixturePath string) {
-	wd, err := os.Getwd()
-	s.Require().NoError(err)
-	cgroupPath := filepath.Join(s.dir, pidCgroupPath)
-	s.Require().NoError(os.MkdirAll(filepath.Dir(cgroupPath), 0755))
-	os.Remove(cgroupPath)
-	s.Require().NoError(os.Symlink(filepath.Join(wd, fixturePath), cgroupPath))
-}
-
-func TestGetContainerIDFromCGroups(t *testing.T) {
-	makeCGroups := func(groupPaths []string) []cgroups.Cgroup {
-		var out []cgroups.Cgroup
-		for _, groupPath := range groupPaths {
-			out = append(out, cgroups.Cgroup{
-				GroupPath: groupPath,
-			})
-		}
-		return out
-	}
-
-	for _, tt := range []struct {
-		name              string
-		cgroupPaths       []string
-		expectPodUID      types.UID
-		expectContainerID string
-		expectCode        codes.Code
-		expectMsg         string
-	}{
-		{
-			name:              "no cgroups",
-			cgroupPaths:       []string{},
-			expectPodUID:      "",
-			expectContainerID: "",
-			expectCode:        codes.OK,
-		},
-		{
-			name: "no container ID in cgroups",
-			cgroupPaths: []string{
-				"/user.slice",
-			},
-			expectPodUID:      "",
-			expectContainerID: "",
-			expectCode:        codes.OK,
-		},
-		{
-			name: "one container ID in cgroups",
-			cgroupPaths: []string{
-				"/user.slice",
-				"/kubepods/pod2c48913c-b29f-11e7-9350-020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			},
-			expectPodUID:      "2c48913c-b29f-11e7-9350-020968147796",
-			expectContainerID: "9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			expectCode:        codes.OK,
-		},
-		{
-			name: "pod UID canonicalized",
-			cgroupPaths: []string{
-				"/user.slice",
-				"/kubepods/pod2c48913c_b29f_11e7_9350_020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			},
-			expectPodUID:      "2c48913c-b29f-11e7-9350-020968147796",
-			expectContainerID: "9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			expectCode:        codes.OK,
-		},
-		{
-			name: "more than one container ID in cgroups",
-			cgroupPaths: []string{
-				"/user.slice",
-				"/kubepods/pod2c48913c-b29f-11e7-9350-020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-				"/kubepods/kubepods/besteffort/pod2c48913c-b29f-11e7-9350-020968147796/a55d9ac3b312d8a2627824b6d6dd8af66fbec439bf4e0ec22d6d9945ad337a38",
-			},
-			expectPodUID:      "",
-			expectContainerID: "",
-			expectCode:        codes.FailedPrecondition,
-			expectMsg:         "multiple container IDs found in cgroups (9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961, a55d9ac3b312d8a2627824b6d6dd8af66fbec439bf4e0ec22d6d9945ad337a38)",
-		},
-		{
-			name: "more than one pod UID in cgroups",
-			cgroupPaths: []string{
-				"/user.slice",
-				"/kubepods/pod11111111-b29f-11e7-9350-020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-				"/kubepods/kubepods/besteffort/pod22222222-b29f-11e7-9350-020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			},
-			expectPodUID:      "",
-			expectContainerID: "",
-			expectCode:        codes.FailedPrecondition,
-			expectMsg:         "multiple pod UIDs found in cgroups (11111111-b29f-11e7-9350-020968147796, 22222222-b29f-11e7-9350-020968147796)",
-		},
-	} {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			podUID, containerID, err := getPodUIDAndContainerIDFromCGroups(makeCGroups(tt.cgroupPaths))
-			spiretest.RequireGRPCStatus(t, err, tt.expectCode, tt.expectMsg)
-			if tt.expectCode != codes.OK {
-				assert.Empty(t, containerID)
-				return
-			}
-			assert.Equal(t, tt.expectPodUID, podUID)
-			assert.Equal(t, tt.expectContainerID, containerID)
-		})
-	}
-}
-
-func TestGetPodUIDAndContainerIDFromCGroupPath(t *testing.T) {
-	for _, tt := range []struct {
-		name              string
-		cgroupPath        string
-		expectPodUID      types.UID
-		expectContainerID string
-	}{
-		{
-			name:              "without QOS",
-			cgroupPath:        "/kubepods/pod2c48913c-b29f-11e7-9350-020968147796/9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-			expectPodUID:      "2c48913c-b29f-11e7-9350-020968147796",
-			expectContainerID: "9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-		},
-		{
-			name:              "with QOS",
-			cgroupPath:        "/kubepods/burstable/pod2c48913c-b29f-11e7-9350-020968147796/34a2062fd26c805aa8cf814cdfe479322b791f80afb9ea4db02d50375df14b41",
-			expectPodUID:      "2c48913c-b29f-11e7-9350-020968147796",
-			expectContainerID: "34a2062fd26c805aa8cf814cdfe479322b791f80afb9ea4db02d50375df14b41",
-		},
-		{
-			name:              "docker for desktop with QOS",
-			cgroupPath:        "/kubepods/kubepods/besteffort/pod6bd2a4d3-a55a-4450-b6fd-2a7ecc72c904/a55d9ac3b312d8a2627824b6d6dd8af66fbec439bf4e0ec22d6d9945ad337a38",
-			expectPodUID:      "6bd2a4d3-a55a-4450-b6fd-2a7ecc72c904",
-			expectContainerID: "a55d9ac3b312d8a2627824b6d6dd8af66fbec439bf4e0ec22d6d9945ad337a38",
-		},
-		{
-			name:              "kind with QOS",
-			cgroupPath:        "/docker/93529524695bb00d91c1f6dba692ea8d3550c3b94fb2463af7bc9ec82f992d26/kubepods/besteffort/poda2830d0d-b0f0-4ff0-81b5-0ee4e299cf80/09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-			expectPodUID:      "a2830d0d-b0f0-4ff0-81b5-0ee4e299cf80",
-			expectContainerID: "09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-		},
-		{
-			name:              "systemd with QOS and container runtime",
-			cgroupPath:        "/kubepods.slice/kubepods-burstable.slice/kubepods-burstable-pod2c48913c-b29f-11e7-9350-020968147796.slice/docker-9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961.scope",
-			expectPodUID:      "2c48913c-b29f-11e7-9350-020968147796",
-			expectContainerID: "9bca8d63d5fa610783847915bcff0ecac1273e5b4bed3f6fa1b07350e0135961",
-		},
-		{
-			name:              "from a different cgroup namespace",
-			cgroupPath:        "/../../../burstable/pod095e82d2-713c-467a-a18a-cbb50a075296/6d1234da0f5aa7fa0ccae4c7d2d109929eb9a81694e6357bcd4547ab3985911b",
-			expectPodUID:      "095e82d2-713c-467a-a18a-cbb50a075296",
-			expectContainerID: "6d1234da0f5aa7fa0ccae4c7d2d109929eb9a81694e6357bcd4547ab3985911b",
-		},
-		{
-			name:              "not kubepods",
-			cgroupPath:        "/something/poda2830d0d-b0f0-4ff0-81b5-0ee4e299cf80/09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-			expectPodUID:      "a2830d0d-b0f0-4ff0-81b5-0ee4e299cf80",
-			expectContainerID: "09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-		},
-		{
-			name:              "just pod uid and container",
-			cgroupPath:        "/poda2830d0d-b0f0-4ff0-81b5-0ee4e299cf80/09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-			expectPodUID:      "a2830d0d-b0f0-4ff0-81b5-0ee4e299cf80",
-			expectContainerID: "09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-		},
-		{
-			name:       "just container segment",
-			cgroupPath: "/09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-		},
-		{
-			name:       "no container segment",
-			cgroupPath: "/kubepods/poda2830d0d-b0f0-4ff0-81b5-0ee4e299cf80",
-		},
-		{
-			name:       "no pod uid segment",
-			cgroupPath: "/kubepods/09bc3d7ade839efec32b6bec4ec79d099027a668ddba043083ec21d3c3b8f1e6",
-		},
-		{
-			name:              "cri-containerd",
-			cgroupPath:        "/kubepods-besteffort-pod72f7f152_440c_66ac_9084_e0fc1d8a910c.slice:cri-containerd:b2a102854b4969b2ce98dc329c86b4fb2b06e4ad2cc8da9d8a7578c9cd2004a2",
-			expectPodUID:      "72f7f152-440c-66ac-9084-e0fc1d8a910c",
-			expectContainerID: "b2a102854b4969b2ce98dc329c86b4fb2b06e4ad2cc8da9d8a7578c9cd2004a2",
-		},
-		{
-			name:       "uid generateds by kubernetes",
-			cgroupPath: "/kubepods/pod2732ca68f6358eba7703fb6f82a25c94",
-		},
-	} {
-		tt := tt
-		t.Run(tt.name, func(t *testing.T) {
-			t.Logf("cgroup path=%s", tt.cgroupPath)
-			podUID, containerID, ok := getPodUIDAndContainerIDFromCGroupPath(tt.cgroupPath)
-			if tt.expectContainerID == "" {
-				assert.False(t, ok)
-				assert.Empty(t, podUID)
-				assert.Empty(t, containerID)
-				return
-			}
-			assert.True(t, ok)
-			assert.Equal(t, tt.expectPodUID, podUID)
-			assert.Equal(t, tt.expectContainerID, containerID)
-		})
-	}
 }
 
 type testFS string
